requirements.txt
Flask==2.1.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
python-dotenv==0.19.0
marshmallow==3.13.0
bcrypt==3.2.0
Werkzeug==2.1.0
pytest==7.1.0
pytest-flask==1.2.0
